Rename package Interactions to Input
