+++ src/main/java/Controller/ControllerClient.java
@@ -1,7 +1,7 @@
 package Controller;
 
 import DAO.ClientsDAO;
-import Interaction.InputManager;
+import Input.InputManager;
 import Model.Client;
 import Model.Product;
 import View.AbstractView;
