+++ src/main/java/Input/InputManager.java
@@ -1,4 +1,4 @@
-package Interaction;
+package Input;
 
 import View.TerminalView;
